fixed warnings

+++ libembryo/include/AspectCell.h
@@ -8,30 +8,24 @@
 #define EMBRYO_ASPECTCELL_H
 #include <list>
 
-//using namespace std;
 
 namespace embryo {
-  
-//   typedef AspectCell * PAspectCell;
-//   typedef std::List<PAspectCell> AspectCellContainer;
-//   typedef AspectCellContainer::iterator AspectCellIterator;
-  
   class AspectCell {
     private:
-      int mPositionX;
-      int mPositionY;
-      int mSize;
+      size_t mPositionX;
+      size_t mPositionY;
+      size_t mSize;
       double mAspect;
     public:
       AspectCell();
-      AspectCell(int iPositionX, int iPositionY, int iSize, double iAspect);
+      AspectCell(size_t iPositionX, size_t iPositionY, size_t iSize, double iAspect);
       ~AspectCell();
-      void setCell(int iPositionX, int iPositionY, int iSize, double iAspect);
+      void setCell(size_t iPositionX, size_t iPositionY, size_t iSize, double iAspect);
       inline double getAspect() { return mAspect; }
       inline void setAspect(double iAspect) { mAspect = iAspect; }
-      inline int getPositionX() { return mPositionX; }
-      inline int getPositionY() { return mPositionY; }
-      inline int getSize() { return mSize; }
+      inline size_t getPositionX() { return mPositionX; }
+      inline size_t getPositionY() { return mPositionY; }
+      inline size_t getSize() { return mSize; }
   };
   
    typedef AspectCell * PAspectCell;
